fix: amend existing tests
